로그 제거, processData 이상해서 예제로 바꿈
뭐가 다른거지

+++ demo_server/demo_server/protocol.h
@@ -492,7 +492,7 @@ struct bl_packet_room_ready {
 	BYTE type;
 
 	short id;
-	short room_no;
+	int room_no;
 	bool is_host;
 };
